Update export
separation of data generation and file format generation
