+++ wordpress/wp-content/plugins/bergclub-plugin/Export/Data/AbstractGenerator.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace BergclubPlugin\Export\Data;
+
+
+abstract class AbstractGenerator implements Generator
+{
+    /**
+     * @var array $args;
+     */
+    protected $args;
+
+    public function __construct(array $args = [])
+    {
+        $this->args = $args;
+    }
+}
