+++ wordpress/wp-content/plugins/bergclub-plugin/Export/Data/GeneratorFactory.php
@@ -8,11 +8,11 @@ use BergclubPlugin\MVC\Helpers;
 
 class GeneratorFactory implements Factory
 {
-    public static function getConcrete($dataSlug)
+    public static function getConcrete($dataSlug, array $args = [])
     {
         $className = __NAMESPACE__ . "\\" . Helpers::snakeToCamelCase($dataSlug) . "Generator";
         if(class_exists($className)){
-            return new $className();
+            return new $className($args);
         }
 
         return null;
