+++ wordpress/wp-content/plugins/bergclub-plugin/Export/Factory.php
@@ -4,5 +4,5 @@ namespace BergclubPlugin\Export;
 
 interface Factory
 {
-    public static function getConcrete($type);
+    public static function getConcrete($type, array $args = []);
 }
