+++ wordpress/wp-content/plugins/bergclub-plugin/Export/Format/AbstractFormat.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace BergclubPlugin\Export\Format;
+
+
+abstract class AbstractFormat implements Format
+{
+    /**
+     * @var array $args;
+     */
+    protected $args;
+
+    public function __construct(array $args = [])
+    {
+        $this->args = $args;
+    }
+}
