+++ wordpress/wp-content/plugins/bergclub-plugin/Export/Format/FormatFactory.php
@@ -7,11 +7,11 @@ use BergclubPlugin\MVC\Helpers;
 
 class FormatFactory implements Factory
 {
-    public static function getConcrete($formatSlug)
+    public static function getConcrete($formatSlug, array $args = [])
     {
         $className = __NAMESPACE__ . "\\" . Helpers::snakeToCamelCase($formatSlug) . "Format";
         if(class_exists($className)){
-            return new $className();
+            return new $className($args);
         }
 
         return null;
